Update example to use friendly error messages
Dearest Reviewer,

Right now if you can not load a file it says something like fopen failed. Adding an expects with which file failed would help a beginner out greatly. 

I have also personally changed the output file name to be the {base name of the source}-{base name of the content}-{step_id}  to help me save all my amazing images. That can be another change or an exercise  for the reader.

Thanks again for the amazing example.

Becker

Changes:
Moved from `?` to `expect` with a message when loading the inputed files.