Fix build on windows for PyTorch 1.8.1
On inspection of torch_cuda_cu.lib using
`dumpbin`, the following symbol shows up:

                  ?getCurrentCUDABlasHandle@cuda@at@@YAPEAUcublasContext@@XZ (struct cublasContext * __cdecl at::cuda::getCurrentCUDABlasHandle(void))

The return type is a struct cublasContext* but in dummy_cuda_dependency.cpp it was declared as void which causes a link error on Windows.

I haven't tested this on Linux since I don't have a working CUDA version for WSL. Could someone verify that this still works cross platform?