Use generics instead of macros for From, TryFrom, TryInto impls
In most cases the macros were not necessary and can be replaced by
a generic implementation over `T: crate::kind::Element`.

The `T: Default` bound is to allow zero-initializing Vec in
`From<&Tensor> for Vec<T>`.
There's probably an unsafe way to avoid the zero-initialization since
those values should be overwritten anyway, but that's out of scope for
this commit.