Change Linear::forward() to match pytorch (#666)
* Change Linear::forward() to match pytorch

The current implementation of `Linear::forward()` is valid and correct; however it produces outputs
that are similar but not numerically identical to `torch.nn.Linear` in PyTorch. This commit replaces
the implementation with `Tensor::linear()`, which matches the PyTorch outputs exactly.

A Python script is included which generates a random linear layer and input tensor, then writes the
tensors and their expected output to disk. A Rust unit test then confirms that `tch-rs` produces the
same output. The test is marked `#[ignore]`, because the tensors themselves are not committed (they
would be about 3M, which feels a little large; if it's fine then they can be committed).

* Reduce size of test tensors, so they're small enough to commit.

These are the smallest tensors I could find that reproduce the difference in output between matmul()
and forward(). This commit also removes the `#[ignore]` attribute, because the test tensors are
small enough now that they can be committed to the repo.