Require libstdc++ even when the compiler looks like Clang.
This is required because some Clang-based compilers (e.g. zig c++) do
not use nor support libstdc++ at all, so it has to use GCC anyway. For
convenience, this condition will be automatically detected and fallback
to GCC will be used, which can be overridden via `LIBTORCH_GCC_FALLBACK`
environment variable for cross-compilation.